Updated openstack/openstack Project: openstack-dev/devstack  629ba0630eb606070378e5edf6058377a9260d98
Disable in-band control for PUBLIC_BRIDGE

By default, Open vSwitch installs internal flows for in-band control
to bridges with controller setting.  "with controller setting" part
is false for openvswitch agent but can be true for ofagent.
Unfortunately the internal flows are incompatible with our use of
the local port here to communicate with neutron router.  This commit
avoids the problem by disabling the in-band control functionality.

Related to blueprint ofagent-port-monitor
Change-Id: Ia79d12da3c2224366425be5f6927859181e18909