Updated openstack/openstack Project: openstack/heat  cbcfe1cedf7d8eb541d568d02a1b25125c606b8b
Use new template for resource update reparse

If the stack-update is switching from pre-HOT to HOT then
the update will fail because the functions are resolved
using the pre-HOT template.

This change uses the new template for the reparse instead of the
old one. It fixes the issue observed but any other effects are
unknown.

Change-Id: I2f749375ccd7ca917387ef96856b1fc248c8567b
Closes-Bug: #1356097