Updated openstack/openstack Project: openstack/keystone  a1da397aa85d8b4f9cbce5788c3b7c7f1a058bcc
Do not require method attribute on plugins

Removes the condition that an authentication plugin knows the "method"
name that is going to be used to call it. This condition prevents
different mechanisms like "kerberos" and "saml" from using the same
backend plugin.

The client should not know how the server is enforcing the Kerberos
authentication,  mod_auth_kerb or embedded Kerberos,  but the
mod_auth_kerb implementation needs to use the same implementation as an
X509 implementation.

Closes-Bug: #1343709

Change-Id: I6c7d44d3809e5e88cc50c50b6df6f3a154df7ab2