Updated openstack/openstack Project: openstack/taskflow  1f6e1aa8e35bd0142063cae81a26c25733a3113b
Allow a jobs posted book to be none by default

Not all users of jobs and jobboards are using the
associated book that accompanies a job, instead they
are fine with just using the name and the details
that can be provided for usage in their application.

Example: http://review.openstack.org/#/c/91763/

To allow the optional usage of books with jobs (which
is already supported) by default set the book to none
and allow it to be provided on a as needed basis.

Change-Id: I69e370a733e44c45c62177008838c259fd9c9a7c