Updated openstack/openstack Project: openstack/tempest  fcd64c0fa862ab24c990a85f8fef3d39c4f1b528
Add ironic instance rebuild test

This adds verification of rebuilding an instance and preservation
of the ephemeral partition. A file is created on the ephemeral
partition before the rebuild, then after the rebuild, we verify
that it still exists. We also verify the expected ephemeral
partition mount point and size, which is an ironic-specific
test criteria.

Change-Id: I6a00f1835c359bfd639bf19ba2c4616638695359