Updated openstack/openstack Project: openstack/nova  02f37c5618db5ccb69fa3c69ebe81a5f6e717390
Treat instance like an object in _start_building

The _start_building method gets an instance object
but was treating it like a dict, so this change
uses dot notation on the instance object fields.

Adds a unit test since this method wasn't directly
unit tested before.

Part of blueprint compute-manager-objects-juno

Change-Id: I8d6b23c70e82fbc7740f3d6af1362905b895addb