Updated openstack/openstack Project: openstack/nova  514a661c0a2b41f91a60007a767ce53a3a10f406
Report all objects with hash mismatches in a single go

The object hash test previously stopped when it hit the first object with
a mismatched hash. This meant that if you changed a couple objects, you
had to run the test, fix the first, run the test, fix the next, etc.
This patch collects all the data and then asserts the delta set, if
there were mismatches.

Change-Id: Ie35a4dbd6941a1f31e0cef7ea56fa2cc5e9a8ef0