Updated openstack/openstack Project: openstack/nova  c9f9ca5339867852477eddfdb9d717b57b4e79be
Treat instance like object in _validate_instance_group_policy

The _validate_instance_group_policy method gets passed an
instance object but was treating it like a primitive dict,
so this changes it to use dot notation when accessing
fields on the instance object.

Updates the existing test to pass an instance object.

Part of blueprint compute-manager-objects-juno

Change-Id: Icbe9fa06ad8f1b32a556241ec7b3269b1fcd5cd2