Updated openstack/openstack Project: openstack/keystone  9ffdedb7581aacc279a3c16400895fceecfee045
Expose context to create grant and delete grant

To correctly issue a CADF audit event for a change in role
assignments, we need to expose the context at the manager level.
Note that the driver signatures are *not* changing, just the
manager needs to know the context.

implements bp role-assignment-notifications

Change-Id: I116b185f5d1fc3f9cbb03ffcf3ce64c56a73d969