Updated openstack/openstack Project: openstack/horizon  cd2d7578fa4431faa4738735e3001e084524d015
Not display loading image for shelved instances

If, any of the instances are shelved through CLI
those instances status are displayed as shelved offloaded
with a loading image.
Displaying the loading image for this status are not correct.
Displaying the loading image is removed for
the shelved offloaded status.
This problem is reported and fixed in #1318558 but this fix is
working only for rescue not for shelved offloaded.
Now it is fixed for shelved offloaded also.

Change-Id: I39d1e2fc1428923824b303f0837b25d736dc0e3c
Closes-Bug: #1352140
Co-Authored-By: Mizielski Robert <robert.mizielski@cloudwatt.com>