Updated openstack/openstack Project: openstack/heat  2350d2d0cdb6978dc7fd074b08e90dcac6b62e09
Add "backup" column to stack table

Adds a backup boolean field to the stack record, defaulted to False,
with a migration which sets it to True for existing backup stacks,
which are named stackname* (which is an invalid name for any non-
backup stack).  Also adds the related column to the DB model and the
Stack object, so when we store a backup stack the flag will be set.

blueprint: list-nested
Change-Id: I737596a7248e8dd0c18a3f444f3b39a2b2e966c8