Updated openstack/openstack Project: openstack/neutron  64a4fadde0f8a1a6413ea5623fc30b9872896a67
Remove unused network parameter from _allocate_ips_for_port

I found this in the context of a patch that I was reviewing [1].  I
found that patch left a few loose ends and so I thought it would be
better to remove the parameter as a separate refactor so that it can
be carefully reviewed.  The other patch [1] should be rebased to this
one.

[1] https://review.openstack.org/#/c/100963

Change-Id: I75115fa46ea3340ef0f13feb28ad2cdb0387fed7