Updated openstack/openstack Project: openstack-dev/devstack  f8f97a677a1419a1ae8e1c5887c82bbea240c6bc
Guard for worlddump in case LOGDIR is undefined

If LOGDIR is undefined, we'll end up calling worldump with the `-d` option but
pass in an empty string as the option value. This causes an argument parsing
error inside worlddump.

The proposed fix is to detect if LOGDIR is undefined, and if so, omit the `-d`
argument.

Change-Id: I77968849ebd3e3bebd2e5929b87b60fc1b6541bf
Closes-Bug: 1333321