Updated openstack/openstack Project: openstack-dev/devstack  33dc8699e6b774cc39b39e0dba97be41bbb47abe
Latest tox should work for openstack now, use it

Remove the tox pin that was in place to avoid issues with tox 1.7.0 and
1.7.1. Latest tox (1.7.2) is compatbile with openstack's tox.ini files
again and should work with tempest. This change is self testing and will
confirm this.

Making this change is important so that devs can stop worrying about
package version pins and helps ensure we don't get left behind by
upstream.

Change-Id: Ibf2102d6ad439dd354ae687f7aafec9ceca6cf9c