Updated openstack/openstack Project: openstack/tempest  7a018e9e51de8d4a1be66a60370fe0a7c2942f06
test_hotplug_nic wait for the guest

The hot_plug nic test expect the new nic is visible,
on the guest almost immediately after the server
 associated to a port.

It is not  true, even the guest needs some time to initialize
the interface.

This change gives time:
* to the guest to realize the new interface.
* to nova/libvirt to actually finish the interface plugging.

Also having the call_until_true calls to respect the configured,
timeout settings and using an exception instead of an assert.

Change-Id: I018e6edefdd3a2f3b8c8d408291464cb28bc8e1e
Closes-Bug: #1358634