Updated openstack/openstack Project: openstack/oslo-incubator  effc2ab39bb9f48d09b6d3040e55b372ce9e6789
Make strutils.mask_password more secure

Make some enhancements to strutils.mask_password to allow it to catch
more cases of passwords in strings. Test cases have been added to test
for these newly added situations.

The following is a listing of patterns that will be handled. The
keyword that mask_password uses (a list of four now) is represented by
<key> and the password is shown as <password>. Quotes (both single and
double) are represented as <quote>.

--<key> <password>
--<key> <quote><password><quote>
<key> = <password>
<key> = <quote><password><quote>

All existing tests and patterns are still handled.

Change-Id: If5ea2d91b1d87c995f50d07a1281879493bd7adb
Partial-Bug: #1345233