Updated openstack/openstack Project: openstack/python-ceilometerclient  98127c4e40a37353fe486483ffc9b8ac25c27474
Removed undefined method in install_env.py file

There was a call from install_venv.py file to post_process method
of InstallVenv class, but this method is not defined in this class
as result it was raising an error. Even when this error doesn't
affect dependencies installation, it is not allowing to display user
information.

Change-Id: I39c995adfbb411f21ad7ea222b2392da6946e25b
Closes-Bug: #1275025