Updated openstack/openstack Project: openstack/nova  5f28f7df45ac80a76cd38d7310ed16b4399ad677
Add graphviz to list of distro packages to install

Commit a507d42cf5d9912c2b3622e84afb8b7d3278595b makes the
doc builds fail on warnings, so people are probably going
to be running 'tox -e docs' more often. To run that you
need the graphviz package from the distro, so add it to
the list of packages needed when setting up a development
environment.

Change-Id: I47ffe63fddfbbd73c3fc698490ad07235caaf459
Closes-Bug: #1356983