Updated openstack/openstack Project: openstack/nova  cbec056e15573a3e6ee402b3b2f30232679f20bb
Use rfc3986 library to validate URL paths and URIs

More work needs to be done in rfc3986 to give the user more control over what
they consider to be a valid URI in the context of RFC 3986. For example, a
previous incarnation of these tests checked that "1" and "abc" were invalid
when according to the RFC they are.

Update the API samples and tests to use valid URIs

DocImpact

Change-Id: I288fbaead64990db1053b7a11e82904611b8498f