Updated openstack/openstack Project: openstack/python-heatclient  c29f12a2a8f251573e537684c0b395d3039bbe9a
Display stack owner info on stack-list

When listing stacks using the --global-tenant flag, the user is not able
to tell who each stack belongs to.  Furthermore, the problem may also
happen if there are multiple users within the same project.

This adds a --show-owner flag to stack-list to display the stack_owner
information when doing a normal stack list.  This flag is also set to
true when using --global-tenant, including an extra project field as
well, since the username is not unique across all projects.

Closes-Bug: #1359898
Change-Id: Id517be1d147f0b3a72ef7c3f8e4d16c7460477f8