Updated openstack/openstack Project: openstack/nova  f8af69ab0aa2633406a2403ea6b4acd05c0ade2e
Fixes hyper-v volume attach when host is AD member

The domain name gets added to the initiator name used by the host
if it's an Active Directory member.

Currently the method which gets the initiator name does not take
this into account when retrieving the default initiator name.

Trying to use a wrong initiator name will lead to an exception when
trying to log in to the according iSCSI target.

This patch simply appends the domain name (when the host is an AD
member) to the initiator name used to log in to iSCSI targets.

Closes-Bug: #1328870
(cherry-picked from commit 0d9419f2b36561bf50bbe0f720c8dd8ad4e96660)

Change-Id: Ifbe762c685e46081059a01043431b2c4ac5473fc